ClickhouseShare to models
