+++ nomium/shares-logger/src/models/mod.rs
@@ -1,2 +1,5 @@
 mod share;
+mod clickhouse_share;
+
 pub use share::ShareLog;
+pub use clickhouse_share::ClickhouseShare;
